Implements training for baseline WaveRNN
This diff implements training with a `wavernn train` command. This
command takes a `--path path/to/model/dir` indicating where to create or
save the model and (on first run) the argument `--config path/to/config.yaml`,
pointing to a YAML training config. On first run, it reads the config
and copies it into the training directory; on the second run, it reads
the config directly from the training directory.